Move systray extension out of generated(?) gui.py file
